requirements.txt
streamlit>=1.31.0
langchain>=0.3.19
langchain-google-genai>=2.0.0
langchain-community>=0.3.18
google-generativeai>=0.7.0
tavily-python>=0.3.3
pypdf2>=3.0.1
pandas>=2.2.0
python-dotenv>=1.0.1
pydantic>=2.0.0
